requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.1.11
cryptography==44.0.3
distro==1.9.0
docx2pdf==0.1.8
et_xmlfile==2.0.0
fire==0.7.0
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.58.0
fpdf==1.7.2
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
packaging==25.0
pandas==2.0.3
pdf2docx==0.5.6
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.0
Pillow==10.0.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMuPDF==1.25.5
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-pptx==0.6.21
pytweening==1.2.0
pytz==2025.2
pywin32==310
setuptools==80.7.1
six==1.17.0
tabula-py==2.7.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uuid==1.30
Werkzeug==2.3.7
wheel==0.45.1
xlrd==2.0.1
XlsxWriter==3.2.3
yt-dlp==2025.2.19
